Refactor: Remove & Rename Files
