+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:get/get.dart';
 import 'package:iris_flutter/config/config.dart';
 import 'package:iris_flutter/config/theme.dart';
-import 'package:iris_flutter/view/page/login/login_noFirebase.dart';
+import 'package:iris_flutter/view/page/login/login_page.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -27,7 +27,7 @@ class MyApp extends StatelessWidget {
                     borderRadius: BorderRadius.circular(15.0)))),
         useMaterial3: true,
       ),
-      home: const LoginTest(),
+      home: const LoginPage(),
       getPages: Config.routers,
     );
   }
